% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTExIdConvert.R
\name{GTExIdConvert}
\alias{GTExIdConvert}
\title{GTExIdConverter}
\usage{
GTExIdConvert(variant_ids)
}
\arguments{
\item{variant_ids}{required. A vector of GTEx variant IDs}
}
\description{
This functions allows you to convert GTEx variant IDs to rsIDs.
}
\examples{
\dontrun{
data(gtex_variant_ids)
gtexIdConvert()}

}
\keyword{eQTL}
\keyword{gtex,}
\keyword{variants,}
